Pull `importFn` out of preview
